refactor(BJS2-1605) {   fixed conflicts }
