+++ src/main/java/faang/school/projectservice/mapper/resource/ResourceMapper.java
@@ -3,16 +3,12 @@ package faang.school.projectservice.mapper.resource;
 import faang.school.projectservice.dto.client.ResourceDto;
 import faang.school.projectservice.model.Project;
 import faang.school.projectservice.model.Resource;
-import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.ReportingPolicy;
 
-@Mapper(componentModel = "spring",
-        unmappedTargetPolicy = ReportingPolicy.IGNORE,
-        injectionStrategy = InjectionStrategy.FIELD)
+@Mapper(componentModel = "Spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface ResourceMapper {
-
     @Mapping(source = "project.id", target = "projectId")
     ResourceDto toDto(Resource resource);
 
@@ -27,5 +23,4 @@ public interface ResourceMapper {
                 .id(id)
                 .build();
     }
-
 }
