+++ src/main/java/faang/school/projectservice/service/project/ProjectFilterService.java
@@ -1,10 +1,10 @@
 package faang.school.projectservice.service.project;
 
 import faang.school.projectservice.config.context.UserContext;
-import faang.school.projectservice.dto.project.ProjectDto;
 import faang.school.projectservice.dto.project.ProjectFilterDto;
-import faang.school.projectservice.filter.ProjectFilter;
+import faang.school.projectservice.filter.project.ProjectFilter;
 import faang.school.projectservice.jpa.TeamMemberJpaRepository;
+import faang.school.projectservice.model.Project;
 import faang.school.projectservice.model.ProjectVisibility;
 import faang.school.projectservice.model.TeamMember;
 import lombok.RequiredArgsConstructor;
@@ -21,21 +21,21 @@ public class ProjectFilterService {
     private final UserContext userContext;
     private final TeamMemberJpaRepository teamMemberJpaRepository;
 
-    public Stream<ProjectDto> applyFilters(Stream<ProjectDto> projectDtoStream, ProjectFilterDto filterDto) {
-        projectDtoStream = filterVisibilityScope(projectDtoStream);
+    public Stream<Project> applyFilters(Stream<Project> projectStream, ProjectFilterDto filterDto) {
+        projectStream = filterVisibilityScope(projectStream);
         for (ProjectFilter projectFilter : projectFilters) {
             if (projectFilter.isApplicable(filterDto)) {
-                projectDtoStream = projectFilter.apply(projectDtoStream, filterDto);
+                projectStream = projectFilter.apply(projectStream, filterDto);
             }
         }
-        return projectDtoStream;
+        return projectStream;
     }
 
-    private Stream<ProjectDto> filterVisibilityScope(Stream<ProjectDto> projectDtoStream) {
-        return projectDtoStream.filter(projectDto -> {
-            if (projectDto.getVisibility().equals(ProjectVisibility.PRIVATE)) {
+    private Stream<Project> filterVisibilityScope(Stream<Project> projectStream) {
+        return projectStream.filter(project -> {
+            if (project.getVisibility().equals(ProjectVisibility.PRIVATE)) {
                 long clientUserId = userContext.getUserId();
-                TeamMember teamMember = teamMemberJpaRepository.findByUserIdAndProjectId(clientUserId, projectDto.getId());
+                TeamMember teamMember = teamMemberJpaRepository.findByUserIdAndProjectId(clientUserId, project.getId());
                 return teamMember != null;
             }
             return true;
